backend/requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==4.2.13
django-cors-headers==4.4.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3